# Fix 3893: return 505 if the request http version is not supported