--- a/docs/Reference/Routes.md
+++ b/docs/Reference/Routes.md
@@ -653,3 +653,9 @@ fastify.route({
   }
 })
 ```
+
+### ⚠  HTTP version check
+
+The fasatify server will run a check on the HTTP version of the request on every call.
+Based on the fastify configuration the check is against one or all the following versions: `2.0`, `1.1`, `1.0` .
+These are the only versions supported, if your fastify server receive a different HTTP version in the request, it will return a `505 HTTP Version Not Supported` error.