--- a/docs/Reference/Routes.md
+++ b/docs/Reference/Routes.md
@@ -658,4 +658,4 @@ fastify.route({
 
 The fasatify server will run a check on the HTTP version of the request on every call.
 Based on the fastify configuration the check is against one or all the following versions: `2.0`, `1.1`, `1.0` .
-These are the only versions supported, if your fastify server receive a different HTTP version in the request, it will return a `505 HTTP Version Not Supported` error.
+These are the only versions supported, if your fastify server receives a different HTTP version in the request, it will return a `505 HTTP Version Not Supported` error.