--- a/lib/route.js
+++ b/lib/route.js
@@ -323,6 +323,16 @@ function buildRouting (options) {
 
   // HTTP request entry point, the routing has already been executed
   function routeHandler (req, res, params, context) {
+    if (isHttpVersionInvalid(req, context)) {
+      const headers = {
+        'Content-Type': 'application/json',
+        'Content-Length': '94'
+      }
+      res.writeHead(505, headers)
+      res.end('{"error":"HTTP Version Not Supported","message":"HTTP Version Not Supported","statusCode":505}')
+      return
+    }
+
     if (closing === true) {
       /* istanbul ignore next mac, windows */
       if (req.httpVersionMajor !== 2) {
@@ -502,6 +512,19 @@ function preParsingHookRunner (functions, request, reply, cb) {
  */
 function removeTrackedSocket () {
   this.keepAliveConnections.delete(this.socket)
+}
+
+function isHttpVersionInvalid (req, context) {
+  const { http2, https } = context.server.initialConfig
+  if (http2 === true && https && https.allowHTTP1 === true) {
+    return !['1.0', '1.1', '2.0'].includes(req.httpVersion)
+  }
+
+  if (http2 === true && (!https || https.allowHTTP1 !== true)) {
+    return req.httpVersion !== '2.0'
+  }
+
+  return !['1.0', '1.1'].includes(req.httpVersion)
 }
 
 function noop () { }