--- a/lib/route.js
+++ b/lib/route.js
@@ -323,13 +323,14 @@ function buildRouting (options) {
 
   // HTTP request entry point, the routing has already been executed
   function routeHandler (req, res, params, context) {
-    if (isHttpVersionInvalid(req, context)) {
+    if (!isHttpVersionValid(req, context)) {
+      const message = '{"error":"HTTP Version Not Supported","message":"HTTP Version Not Supported","statusCode":505}'
       const headers = {
         'Content-Type': 'application/json',
-        'Content-Length': '94'
+        'Content-Length': message.length
       }
       res.writeHead(505, headers)
-      res.end('{"error":"HTTP Version Not Supported","message":"HTTP Version Not Supported","statusCode":505}')
+      res.end(message)
       return
     }
 
@@ -341,12 +342,13 @@ function buildRouting (options) {
       }
 
       if (return503OnClosing) {
+        const message = '{"error":"Service Unavailable","message":"Service Unavailable","statusCode":503}'
         const headers = {
           'Content-Type': 'application/json',
-          'Content-Length': '80'
+          'Content-Length': message.length
         }
         res.writeHead(503, headers)
-        res.end('{"error":"Service Unavailable","message":"Service Unavailable","statusCode":503}')
+        res.end(message)
         return
       }
     }
@@ -514,17 +516,17 @@ function removeTrackedSocket () {
   this.keepAliveConnections.delete(this.socket)
 }
 
-function isHttpVersionInvalid (req, context) {
+function isHttpVersionValid (req, context) {
   const { http2, https } = context.server.initialConfig
   if (http2 === true && https && https.allowHTTP1 === true) {
-    return !['1.0', '1.1', '2.0'].includes(req.httpVersion)
+    return ['1.0', '1.1', '2.0'].includes(req.httpVersion)
   }
 
   if (http2 === true && (!https || https.allowHTTP1 !== true)) {
-    return req.httpVersion !== '2.0'
+    return req.httpVersion === '2.0'
   }
 
-  return !['1.0', '1.1'].includes(req.httpVersion)
+  return ['1.0', '1.1'].includes(req.httpVersion)
 }
 
 function noop () { }