This PR adds a check on the HTTP version when a call is made to the server.
This check is run before the handler is called but after the routing has been executed [1]. The check is based on the initial config for the fastify server and the request http version.

The assumption is to consider `1.0`, `1.1` and `2.0` the only valid HTTP versions. 
This PR adds also some documentation in the routes section [2].

This PR should be fixing both the issues raised in #3893 . The 404 handler will not be called if there is an HTTP version mismatch.

The 505 response will be a json response [3]

```
{"error":"HTTP Version Not Supported","message":"HTTP Version Not Supported","statusCode":505}
```

Notes:
[1] I'm not 100% sure if running the check on the http version would be better placed in the [`preRouting`](https://github.com/fastify/fastify/blob/main/fastify.js#L706) function
[2] Is there a better place in the documentation?
[3] Should it be another content type?

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)