@mcollina 2022-05-18
No if everything works correctly it's automatically backported.