@RafaelGSS 2022-05-18
It seems to have a regression, @p16 would you mind running the benchmarks locally in your machine?