@p16 2022-05-18
@RafaelGSS this branch does seem to be slower 

on master

```
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1]
[1]
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
[1] │ Latency │ 10 ms │ 21 ms │ 24 ms │ 27 ms │ 16.94 ms │ 5.96 ms │ 104 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
[1] ┌───────────┬────────┬────────┬─────────┬─────────┬──────────┬─────────┬────────┐
[1] │ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg      │ Stdev   │ Min    │
[1] ├───────────┼────────┼────────┼─────────┼─────────┼──────────┼─────────┼────────┤
[1] │ Req/Sec   │ 48383  │ 48383  │ 57855   │ 59839   │ 57339.74 │ 2222.59 │ 48383  │
[1] ├───────────┼────────┼────────┼─────────┼─────────┼──────────┼─────────┼────────┤
[1] │ Bytes/Sec │ 9.1 MB │ 9.1 MB │ 10.9 MB │ 11.3 MB │ 10.8 MB  │ 418 kB  │ 9.1 MB │
[1] └───────────┴────────┴────────┴─────────┴─────────┴──────────┴─────────┴────────┘
[1]
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1]
[1] 1721k requests in 30.03s, 323 MB read
[1] npx autocannon -c 100 -d 30 -p 10 localhost:3000/ exited with code 0
--> Sending SIGTERM to other processes..
[0] node ./examples/benchmark/simple.js exited with code SIGTERM
```

this branch

```
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1]
[1]
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
[1] │ Latency │ 10 ms │ 21 ms │ 30 ms │ 35 ms │ 18.61 ms │ 7.16 ms │ 141 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬───────┬──────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5% │ Avg      │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼───────┼──────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 38783   │ 38783   │ 52479   │ 58687 │ 52337.07 │ 3920.31 │ 38762   │
[1] ├───────────┼─────────┼─────────┼─────────┼───────┼──────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 7.29 MB │ 7.29 MB │ 9.86 MB │ 11 MB │ 9.84 MB  │ 737 kB  │ 7.29 MB │
[1] └───────────┴─────────┴─────────┴─────────┴───────┴──────────┴─────────┴─────────┘
[1]
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1]
[1] 1571k requests in 30.03s, 295 MB read
[1] npx autocannon -c 100 -d 30 -p 10 localhost:3000/ exited with code 0
--> Sending SIGTERM to other processes..
[0] node ./examples/benchmark/simple.js exited with code SIGTERM
```