@climba03003 2022-05-18
If regression is noticeable and cannot be fixed.

I recommend to provide a option to disable the check.
The main reason is that the `http version` can be check in reverse proxy that maybe more performant than checking inside fastify. And if your application is place behind a reverse proxy like `nginx`, it should not require a check since it should be always correct.