@RafaelGSS 2022-05-19
@p16 this should fix the perf regression:

```diff
diff --git a/lib/route.js b/lib/route.js
index dbcd0939..eeec9c0a 100644
--- a/lib/route.js
+++ b/lib/route.js
@@ -518,15 +518,17 @@ function removeTrackedSocket () {

 function isHttpVersionValid (req, context) {
   const { http2, https } = context.server.initialConfig
-  if (http2 === true && https && https.allowHTTP1 === true) {
-    return ['1.0', '1.1', '2.0'].includes(req.httpVersion)
-  }
+  if (http2 === true) {
+    if (https && https.allowHTTP1 === true) {
+      return req.httpVersion === '2.0' || req.httpVersion === '1.1' || req.httpVersion === '1.0'
+    }

-  if (http2 === true && (!https || https.allowHTTP1 !== true)) {
-    return req.httpVersion === '2.0'
+    if (!https || https.allowHTTP1 !== true) {
+      return req.httpVersion === '2.0'
+    }
   }

-  return ['1.0', '1.1'].includes(req.httpVersion)
+  return req.httpVersion === '1.1' || req.httpVersion === '1.0'
 }

 function noop () { }
```

The reason is that 'Array.prototype.includes` is very slow in v18. See: https://github.com/RafaelGSS/nodejs-bench-operations/blob/main/RESULTS-v18.md#includes-vs-raw-comparissonjs.

For comparison, see the v16: https://github.com/RafaelGSS/nodejs-bench-operations/blob/main/RESULTS-v16.md#includes-vs-raw-comparissonjs