@Fdawgs 2022-05-19
@RafaelGSS any idea why performance has dipped for includes() in node 18? Would moving to Array.foreach, and putting the comparison in the loop, be any better?