@p16 2022-05-19
@RafaelGSS it improves but I'm not sure it is enough, on average there are still 3000 calls less for this branch VS master

This branch

```
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1]
[1]
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
[1] │ Latency │ 11 ms │ 24 ms │ 35 ms │ 40 ms │ 21.12 ms │ 7.74 ms │ 121 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬──────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg      │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 39455   │ 39455   │ 46431   │ 50271   │ 46260.27 │ 2828.74 │ 39454   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 7.42 MB │ 7.42 MB │ 8.73 MB │ 9.45 MB │ 8.7 MB   │ 531 kB  │ 7.42 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴──────────┴─────────┴─────────┘
[1]
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1]
[1] 1389k requests in 30.03s, 261 MB read
[1] npx autocannon -c 100 -d 30 -p 10 localhost:3000/ exited with code 0
--> Sending SIGTERM to other processes..
[0] node ./examples/benchmark/simple.js exited with code SIGTERM
```

master

```
[1] Running 30s test @ http://localhost:3000/
[1] 100 connections with 10 pipelining factor
[1]
[1]
[1] ┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
[1] │ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
[1] ├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
[1] │ Latency │ 11 ms │ 23 ms │ 32 ms │ 36 ms │ 19.82 ms │ 7.61 ms │ 153 ms │
[1] └─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
[1] ┌───────────┬─────────┬─────────┬─────────┬─────────┬──────────┬─────────┬─────────┐
[1] │ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg      │ Stdev   │ Min     │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┼─────────┤
[1] │ Req/Sec   │ 38047   │ 38047   │ 49535   │ 52607   │ 49233.07 │ 3015.25 │ 38016   │
[1] ├───────────┼─────────┼─────────┼─────────┼─────────┼──────────┼─────────┼─────────┤
[1] │ Bytes/Sec │ 7.15 MB │ 7.15 MB │ 9.31 MB │ 9.89 MB │ 9.26 MB  │ 568 kB  │ 7.15 MB │
[1] └───────────┴─────────┴─────────┴─────────┴─────────┴──────────┴─────────┴─────────┘
[1]
[1] Req/Bytes counts sampled once per second.
[1] # of samples: 30
[1]
[1] 1478k requests in 30.04s, 278 MB read
[1] npx autocannon -c 100 -d 30 -p 10 localhost:3000/ exited with code 0
--> Sending SIGTERM to other processes..
[0] node ./examples/benchmark/simple.js exited with code SIGTERM
```

Should we just drop this or place it behind a flag?